batch: Move merge candidate enumeration
The batch merge path applies selected ownership to normalized source and
target line sequences. It also exposes candidate previews for ambiguous
placements.

The preview path still lives inside the merge orchestration module. That
couples candidate construction to line application flow, even though
callers only need merge to validate a proposed resolution.

This commit addresses that split by moving candidate enumeration into a
merge_candidate_enumeration module. Merge now supplies a validation
callback for proposed resolutions instead of letting the enumerator call
private merge helpers.

A following tests commit will pin the new ownership boundary and update the
architecture expectations for the helper modules consumed by enumeration.
